+++ templates/preview/previewApi.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"net/http"
 	"path"
+	"strings"
 
 	"github.com/gorilla/mux"
 
@@ -164,6 +165,8 @@ func (a *Api) generateEmail(templateName models.TemplateName, lang string) (stri
 
 	log.Printf("trying preview with template '%s' with language '%s'", templateName, lang)
 
+	supportEmail := fmt.Sprintf("<a href=%s>%s</a>", a.Config.SupportURL, strings.Replace(a.Config.SupportURL, "mailto:", "", 1))
+
 	content := map[string]interface{}{
 		"Key":          "123456789123456789123456789123456789",
 		"Email":        "john@diabeloop.com",
@@ -172,6 +175,7 @@ func (a *Api) generateEmail(templateName models.TemplateName, lang string) (stri
 		"WebPath":      "login",
 		"ShortKey":     "12345678",
 		"OTP":          "165236984",
+		"SupportEmail": supportEmail,
 	}
 	// Content collection is here to replace placeholders in template body/content
 	content["CreatorName"] = "John Doe"